feat: consolidate iOS and macOS file picker implementations into a single FilePickerPlugin class
